fetching data from mongodb
