+++ DashX-Backend/env.json
@@ -1,0 +1,6 @@
+{
+    "dev" : {
+        "mongo_username" : "admin",
+        "mongo_password" : "dashx12345" 
+    }
+}
