+++ DashX-Backend/package.json
@@ -16,6 +16,7 @@
     "express-winston": "^4.1.0",
     "google-spreadsheet": "^3.1.15",
     "http": "0.0.1-security",
+    "mongoose": "^6.0.6",
     "nodemon": "^2.0.12",
     "open": "^8.2.1",
     "server-destroy": "^1.0.1",
